app/src/main/java/mega/privacy/android/app/presentation/node/view/bottomsheetmenuitems/SlideshowBottomSheetMenuItem.kt
package mega.privacy.android.app.presentation.node.view.bottomsheetmenuitems

import mega.privacy.android.app.presentation.node.model.menuaction.SlideshowMenuAction
import mega.privacy.android.core.ui.model.MenuActionWithIcon
import mega.privacy.android.domain.entity.node.TypedNode
import mega.privacy.android.domain.entity.shares.AccessPermission
import javax.inject.Inject

/**
 * Slideshow bottom sheet menu item
 *
 * @param menuAction [SlideshowMenuAction]
 */
class SlideshowBottomSheetMenuItem @Inject constructor(
    override val menuAction: SlideshowMenuAction,
) : NodeBottomSheetMenuItem<MenuActionWithIcon> {
    override suspend fun shouldDisplay(
        isNodeInRubbish: Boolean,
        accessPermission: AccessPermission?,
        isInBackups: Boolean,
        node: TypedNode,
        isConnected: Boolean,
    ) = false //Due to requirement change, now just hide slideshow entry in node context menu

    override val groupId = 5
}